chore: refactor onRequestPaused
