+++ src/bidiMapper/domains/network/NetworkRequest.ts
@@ -25,7 +25,7 @@ import type Protocol from 'devtools-protocol';
 import {Deferred} from '../../../utils/Deferred.js';
 import type {EventManager} from '../events/EventManager.js';
 import {
-  type Network,
+  Network,
   type BrowsingContext,
   ChromiumBidi,
 } from '../../../protocol/protocol.js';
@@ -35,8 +35,10 @@ import type {CdpTarget} from '../context/CdpTarget.js';
 
 import {
   computeResponseHeadersSize,
+  bidiNetworkHeadersFromCdpFetchHeaderEntryArray,
   bidiNetworkHeadersFromCdpNetworkHeaders,
 } from './NetworkUtils.js';
+import type {NetworkStorage} from './NetworkStorage.js';
 
 /** Abstracts one individual network request. */
 export class NetworkRequest {
@@ -203,7 +205,65 @@ export class NetworkRequest {
   }
 
   /** Fired whenever a network request interception is hit. */
-  onRequestPaused(_event: Protocol.Fetch.RequestPausedEvent) {
+  onRequestPaused(
+    params: Protocol.Fetch.RequestPausedEvent,
+    networkStorage: NetworkStorage
+  ) {
+    // The stage of the request can be determined by presence of
+    // responseErrorReason and responseStatusCode -- the request is at
+    // the response stage if either of these fields is present and in the
+    // request stage otherwise.
+    let phase: Network.InterceptPhase;
+    if (
+      params.responseErrorReason === undefined &&
+      params.responseStatusCode === undefined
+    ) {
+      phase = Network.InterceptPhase.BeforeRequestSent;
+    } else if (
+      params.responseStatusCode === 401 &&
+      params.responseStatusText === 'Unauthorized'
+    ) {
+      // https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/401
+      phase = Network.InterceptPhase.AuthRequired;
+    } else {
+      phase = Network.InterceptPhase.ResponseStarted;
+    }
+
+    const headers = bidiNetworkHeadersFromCdpFetchHeaderEntryArray(
+      // TODO: Use params.request.headers if request?
+      params.responseHeaders
+    );
+
+    assert(this.requestId === params.networkId);
+    networkStorage.addBlockedRequest(this.requestId, {
+      request: params.requestId, // intercept request id
+      phase,
+      // TODO: Finish populating response / ResponseData.
+      response: {
+        url: params.request.url,
+        // TODO: populate.
+        protocol: '',
+        status: params.responseStatusCode ?? 0,
+        statusText: params.responseStatusText ?? '',
+        // TODO: populate.
+        fromCache: false,
+        headers,
+        // TODO: populate.
+        mimeType: '',
+        // TODO: populate.
+        bytesReceived: 0,
+        headersSize: computeResponseHeadersSize(headers),
+        // TODO: consider removing from spec.
+        bodySize: 0,
+        // TODO: consider removing from spec.
+        content: {
+          size: 0,
+        },
+        // TODO: populate.
+        authChallenge: undefined,
+      },
+    });
+
     this.#isBlocked = true;
   }
